feat(dataobj): Replace poll loop with config-driven flip-flop workers
