+++ pkg/engine/compactor/config.go
@@ -28,16 +28,15 @@ type Config struct {
 	// entry. Empty means no workers run.
 	CompactionTenants flagext.StringSliceCSV `yaml:"tenants"`
 
-	// MaxRunningCompactionTasks caps how many IndexMerge tasks the
-	// coordinator runs concurrently per tenant within a single cycle.
-	// Applied via errgroup.SetLimit on the per-tenant goroutine group in
-	// runTenantCycle. Zero means unlimited (one goroutine per task with no
-	// admission throttle). Negative values are rejected at config validation.
+	// MaxRunningCompactionTasks caps how many IndexMerge tasks the coordinator
+	// runs concurrently per tenant. Zero means unlimited (one goroutine per task
+	// with no admission throttle). Negative values are rejected at config
+	// validation.
 	MaxRunningCompactionTasks int `yaml:"max_running_compaction_tasks"`
 
 	// LogMaxRunningCompactionTasks caps how many LogMerge tasks the coordinator
-	// runs concurrently per tenant within a single cycle. Zero means unlimited.
-	// Negative values are rejected at config validation.
+	// runs concurrently per tenant. Zero means unlimited. Negative values are
+	// rejected at config validation.
 	LogMaxRunningCompactionTasks int `yaml:"logs_max_running_compaction_tasks"`
 
 	// PollingInterval is the cadence of the coordinator's main loop. Each
